+++ Runtime/Properties/InputActionMapPopup.cs
@@ -1,3 +1,4 @@
+using System;
 using UnityEngine.InputSystem;
 
 namespace ActionCode.InputSystem
@@ -5,7 +6,7 @@ namespace ActionCode.InputSystem
     /// <summary>
     /// Displays the action maps from a given <see cref="InputActionAsset"/> as a Popup field.
     /// </summary>
-    [System.Serializable]
+    [Serializable]
     public class InputActionMapPopup
     {
         public string mapName;
@@ -20,25 +21,6 @@ namespace ActionCode.InputSystem
         {
             this.mapName = mapName;
             this.assetName = assetName;
-        }
-
-        /// <summary>
-        /// Finds an ActionMap inside the given actionAsset and returns it.
-        /// </summary>
-        /// <param name="actionAsset">InputActionAsset instance.</param>
-        /// <param name="name">Name of the ActionMap.</param>
-        /// <returns>An InputActionMap instance if found. Null otherwise.</returns>
-        public InputActionMap FindActionMap(InputActionAsset actionAsset, string name)
-        {
-            InputActionMap actionMap = actionAsset.FindActionMap(name);
-            if (actionMap == null)
-            {
-                UnityEngine.Debug.LogErrorFormat("Cannot find action map '{0}' in actions '{1}'", name, actionAsset);
-                return null;
-            }
-
-            mapName = name;
-            return actionMap;
         }
     }
 }